Initial commit (but not the first)
It is the initial commit of the EV3 code belongs to Töviscsapat NextGenPost. However, it is not the real initial commit, just GitHub messed it up a bit, and I had to create a new repo. 
First, I made our MyBlocks, like AbsDRV, AbsTU, MedMotor, Align and LFOL. Now they seem to work properly.

Then, I got to MasterMenu which wasn't that hard but when I just got to the interception of a running round. I've worked with unknown things like multiprocessing or multithreading, and I learned a lot about them, but I think these methods are just too complicated for a computer like EV3.  So I moved on with saving the current round in a text file, so if we quit it can memorize the last round and start the MasterMenu from that. I tried it with DataLogging, but it is for another purpose, so I tried pure text file writing/reading again, and it works. Now I just need to connect it to the exit button.